pip-requirements.txt
Markdown==3.3.7
python-markdown-math==0.8
pydantic==1.9.1
python-dateutil==2.8.2
slugify==0.0.1